feat: extend max object size to 8kB
- Change max object size to 8kB by extending chaning the object_size
  header entry to uint16_t.
- Remove 6th obis code element (F) from ObisCode, assume it is 255 when
  parsing (ignore if not there, ignore if 255, fail otherwise).

+++ components/efs/header.h
@@ -10,7 +10,7 @@ struct Header {
  public:
   ObisCode obis_code;
   uint8_t num_values;
-  uint8_t object_size;  // Size of the current obj. including Header
+  uint16_t object_size;  // Size of the current obj. including Header
 };
 const size_t HEADER_SIZE = sizeof(Header);
 static_assert(HEADER_SIZE == 8, "Header size should be 8 bytes.");
